소개/개발자 페이지 분류, 사용 언어 추가
