+++ src/component/footer.js
@@ -7,9 +7,7 @@ const Footer = () => {
       <div className="footer-container">
         <div className="footer-section">
           <h4>또박또박</h4>
-          <p>
-            중부대학교 정보호호학전공 - 박유진 신윤호 이정호 최하은
-          </p>
+          <p>중부대학교 정보호호학전공 - 박유진 신윤호 이정호 최하은</p>
         </div>
         <div className="footer-section">
           <h4>사이트 맵</h4>
@@ -17,7 +15,9 @@ const Footer = () => {
             <li>
               <a href="/">홈</a>
             </li>
-
+            <li>
+              <a href="/introduce">소개</a>
+            </li>
             <li>
               <a href="/learn">학습하기</a>
             </li>
@@ -31,7 +31,7 @@ const Footer = () => {
               <a href="/notebook">연습장</a>
             </li>
             <li>
-              <a href="/introduce">개발자</a>
+              <a href="/teamIntro">개발자</a>
             </li>
           </ul>
         </div>
