+++ src/css/introduce.css
@@ -36,11 +36,6 @@
   font-size: 1.5em;
 }
 
-.team-member p {
-  margin: 5px 0 0;
-  color: #555;
-}
-
 .site-introduce {
   font-size: 18px;
   line-height: 1.6;
@@ -88,4 +83,8 @@
 .team-member-photo {
   width: 100px; /* 팀원 사진 너비 */
   height: auto; /* 비율에 맞게 높이 자동 조절 */
+}
+
+.progrom-img {
+  width: 900px;
 }
