+++ src/page/introduce.js
@@ -1,37 +1,12 @@
 // introduce.js
 import React from "react";
 import "../css/introduce.css";
-import image1 from "../img/woman.png";
-import image2 from "../img/man.png";
 import introImg from "../img/intro.png";
-import css from "../img/css.png";
-import js from "../img/js.png";
-import nodejs from "../img/nodejs.png";
-import react from "../img/react.png";
-import vscode from "../img/vscode.png";
-
-const teamMembers = [
-  { image: image1, name: "박유진", role: "프론트엔드" },
-  { image: image2, name: "신윤호", role: "백엔드" },
-  { image: image2, name: "이정호", role: "프론트엔드" },
-  { image: image1, name: "최하은", role: "프론트엔드" },
-];
-
-const TeamMember = ({ image, name, role }) => (
-  <div className="team-member">
-    <img src={image} alt={name} className="team-member-photo" />
-    <h2>{name}</h2>
-    <p>{role}</p>
-  </div>
-);
+import programIntro from "../img/programIntro.png";
 
 const TechnologyIcons = () => (
-  <div className="technology-icons">
-    <img src={css} alt="CSS" className="tech-icon" />
-    <img src={js} alt="JavaScript" className="tech-icon" />
-    <img src={nodejs} alt="Node.js" className="tech-icon" />
-    <img src={react} alt="React" className="tech-icon" />
-    <img src={vscode} alt="VSCode" className="tech-icon" />
+  <div>
+    <img src={programIntro} alt="programIntro" className="progrom-img" />
   </div>
 );
 
@@ -60,16 +35,8 @@ const Introduce = () => (
       </div>
     </div>
     <div className="introduceDiv">
-      <h1>팀원 소개</h1>
-      <div className="team-members">
-        {teamMembers.map((member, index) => (
-          <TeamMember key={index} {...member} />
-        ))}
-      </div>
-    </div>
-    <div className="introduceDiv">
-      <h1>사용한 기술, 라이브러리</h1>
-      <div className='site-introduce'>
+      <h1>개발 언어</h1>
+      <div>
         <TechnologyIcons />
       </div>
     </div>
@@ -88,7 +55,7 @@ const Introduce = () => (
         주소를 확인하실 수 있습니다.
       </div>
     </div>
-  </div >
+  </div>
 );
 
 export default Introduce;
